creates new file for GameRunnerSpy and methods from separate classes into it
