+++ src/test/java/de/rabea/gui/clickhandler/ReplayClickHandlerTest.java
@@ -1,6 +1,5 @@
 package de.rabea.gui.clickhandler;
 
-import de.rabea.game.GameRunner;
 import org.junit.Test;
 
 import static org.junit.Assert.assertTrue;
@@ -15,18 +14,4 @@ public class ReplayClickHandlerTest {
 
         assertTrue(guiAppSpy.displayGameModeOptionsWasCalled);
     }
-
-    private class GameRunnerSpy extends GameRunner {
-        private boolean displayGameModeOptionsWasCalled;
-
-        public GameRunnerSpy() {
-            super(null, null);
-        }
-
-        @Override
-        public void displayGameModeOptions() {
-            displayGameModeOptionsWasCalled = true;
-        }
-    }
-
 }
